Cleared types, added asserts for debug of type conversion, removed unused function, added "using" keyword for "protected" override on functions.
Might have broken some EC on the wx callbacks, we'll see. Please test.

+++ trunk/src/libs/ec/cpp/ECSpecialTags.cpp
@@ -148,7 +148,7 @@ wxString CEC_StatTree_Node_Tag::GetDisplayString() const
 		}
 	}
 	CFormat label(my_label);
-	for (int i = 0; i < GetTagCount(); ++i) {
+	for (size_t i = 0; i < GetTagCount(); ++i) {
 		const CECTag *tmp = GetTagByIndex(i);
 		if (tmp->GetTagName() == EC_TAG_STAT_NODE_VALUE) {
 			FormatValue(label, tmp);
